refactor: fixes linting issues
