--- tests/mock_debug.log
@@ -1,48 +1,0 @@
-Array
-(
-    [method] => POST
-    [content_type] => multipart/form-data; boundary=------------------------df44dddb8b55097c
-    [_POST] => Array
-        (
-            [name] => Mark
-        )
-
-    [_FILES] => Array
-        (
-            [files] => Array
-                (
-                    [name] => Array
-                        (
-                            [owl.gif] => upload.txt
-                        )
-
-                    [full_path] => Array
-                        (
-                            [owl.gif] => C:\Users\Hamza\Desktop\Repos\unirest-php\tests/Mocking/upload.txt
-                        )
-
-                    [type] => Array
-                        (
-                            [owl.gif] => application/octet-stream
-                        )
-
-                    [tmp_name] => Array
-                        (
-                            [owl.gif] => C:\Users\Hamza\AppData\Local\Temp\phpA637.tmp
-                        )
-
-                    [error] => Array
-                        (
-                            [owl.gif] => 0
-                        )
-
-                    [size] => Array
-                        (
-                            [owl.gif] => 15
-                        )
-
-                )
-
-        )
-
-)
