--- tests/server.log
@@ -1,88 +1,0 @@
-[Fri Jul  4 12:24:31 2025] PHP 8.2.4 Development Server (http://localhost:8000) started
-[Fri Jul  4 12:24:32 2025] [::1]:52299 Accepted
-[Fri Jul  4 12:24:32 2025] [::1]:52299 Closed without sending a request; it was probably just an unused speculative preconnection
-[Fri Jul  4 12:24:32 2025] [::1]:52299 Closing
-[Fri Jul  4 12:24:33 2025] [::1]:52301 Accepted
-[Fri Jul  4 12:24:33 2025] [::1]:52301 Closing
-[Fri Jul  4 12:24:33 2025] [::1]:52303 Accepted
-[Fri Jul  4 12:24:35 2025] [::1]:52303 Closing
-[Fri Jul  4 12:24:35 2025] [::1]:52305 Accepted
-[Fri Jul  4 12:24:35 2025] [::1]:52305 Closing
-[Fri Jul  4 12:24:35 2025] [::1]:52308 Accepted
-[Fri Jul  4 12:24:35 2025] [::1]:52308 Closing
-[Fri Jul  4 12:24:35 2025] [::1]:52310 Accepted
-[Fri Jul  4 12:24:35 2025] [::1]:52310 Closing
-[Fri Jul  4 12:24:36 2025] [::1]:52312 Accepted
-[Fri Jul  4 12:24:36 2025] [::1]:52312 Closing
-[Fri Jul  4 12:24:36 2025] [::1]:52314 Accepted
-[Fri Jul  4 12:24:36 2025] [::1]:52314 Closing
-[Fri Jul  4 12:24:36 2025] [::1]:52316 Accepted
-[Fri Jul  4 12:24:36 2025] [::1]:52316 Closing
-[Fri Jul  4 12:24:36 2025] [::1]:52318 Accepted
-[Fri Jul  4 12:24:36 2025] [::1]:52318 Closing
-[Fri Jul  4 12:24:36 2025] [::1]:52320 Accepted
-[Fri Jul  4 12:24:36 2025] [::1]:52320 Closing
-[Fri Jul  4 12:24:37 2025] [::1]:52322 Accepted
-[Fri Jul  4 12:24:37 2025] [::1]:52322 Closing
-[Fri Jul  4 12:24:37 2025] [::1]:52324 Accepted
-[Fri Jul  4 12:24:37 2025] [::1]:52324 Closing
-[Fri Jul  4 12:24:37 2025] [::1]:52326 Accepted
-[Fri Jul  4 12:24:37 2025] [::1]:52326 Closing
-[Fri Jul  4 12:24:37 2025] [::1]:52328 Accepted
-[Fri Jul  4 12:24:37 2025] [::1]:52328 Closing
-[Fri Jul  4 12:24:37 2025] [::1]:52330 Accepted
-[Fri Jul  4 12:24:37 2025] [::1]:52330 Closing
-[Fri Jul  4 12:24:38 2025] [::1]:52332 Accepted
-[Fri Jul  4 12:24:38 2025] [::1]:52332 Closing
-[Fri Jul  4 12:24:38 2025] [::1]:52334 Accepted
-[Fri Jul  4 12:24:38 2025] [::1]:52334 Closing
-[Fri Jul  4 12:24:38 2025] [::1]:52336 Accepted
-[Fri Jul  4 12:24:38 2025] [::1]:52336 Closing
-[Fri Jul  4 12:24:38 2025] [::1]:52338 Accepted
-[Fri Jul  4 12:24:38 2025] [::1]:52338 Closing
-[Fri Jul  4 12:24:39 2025] [::1]:52340 Accepted
-[Fri Jul  4 12:24:39 2025] [::1]:52340 Closing
-[Fri Jul  4 12:24:39 2025] [::1]:52342 Accepted
-[Fri Jul  4 12:24:39 2025] [::1]:52342 Closing
-[Fri Jul  4 12:24:39 2025] [::1]:52344 Accepted
-[Fri Jul  4 12:24:39 2025] [::1]:52344 Closing
-[Fri Jul  4 12:24:39 2025] [::1]:52346 Accepted
-[Fri Jul  4 12:24:39 2025] [::1]:52346 Closing
-[Fri Jul  4 12:24:39 2025] [::1]:52348 Accepted
-[Fri Jul  4 12:24:39 2025] [::1]:52348 Closing
-[Fri Jul  4 12:24:40 2025] [::1]:52350 Accepted
-[Fri Jul  4 12:24:40 2025] [::1]:52350 Closing
-[Fri Jul  4 12:24:40 2025] [::1]:52352 Accepted
-[Fri Jul  4 12:24:40 2025] [::1]:52352 Closing
-[Fri Jul  4 12:24:40 2025] [::1]:52354 Accepted
-[Fri Jul  4 12:24:40 2025] [::1]:52354 Closing
-[Fri Jul  4 12:24:40 2025] [::1]:52356 Accepted
-[Fri Jul  4 12:24:40 2025] [::1]:52356 Closing
-[Fri Jul  4 12:24:40 2025] [::1]:52358 Accepted
-[Fri Jul  4 12:24:40 2025] [::1]:52358 Closing
-[Fri Jul  4 12:24:41 2025] [::1]:52360 Accepted
-[Fri Jul  4 12:24:41 2025] [::1]:52360 Closing
-[Fri Jul  4 12:24:41 2025] [::1]:52362 Accepted
-[Fri Jul  4 12:24:41 2025] [::1]:52362 Closing
-[Fri Jul  4 12:24:41 2025] [::1]:52364 Accepted
-[Fri Jul  4 12:24:41 2025] [::1]:52364 Closing
-[Fri Jul  4 12:24:41 2025] [::1]:52366 Accepted
-[Fri Jul  4 12:24:41 2025] [::1]:52366 Closing
-[Fri Jul  4 12:24:42 2025] [::1]:52368 Accepted
-[Fri Jul  4 12:24:42 2025] [::1]:52368 Closing
-[Fri Jul  4 12:24:42 2025] [::1]:52370 Accepted
-[Fri Jul  4 12:24:42 2025] [::1]:52370 Closing
-[Fri Jul  4 12:24:42 2025] [::1]:52372 Accepted
-[Fri Jul  4 12:24:42 2025] [::1]:52372 Closing
-[Fri Jul  4 12:24:42 2025] [::1]:52376 Accepted
-[Fri Jul  4 12:24:42 2025] [::1]:52376 Closing
-[Fri Jul  4 12:24:42 2025] [::1]:52378 Accepted
-[Fri Jul  4 12:24:42 2025] [::1]:52378 Closing
-[Fri Jul  4 12:24:43 2025] [::1]:52380 Accepted
-[Fri Jul  4 12:24:43 2025] [::1]:52380 Closing
-[Fri Jul  4 12:24:43 2025] [::1]:52382 Accepted
-[Fri Jul  4 12:24:43 2025] [::1]:52382 Closing
-[Fri Jul  4 12:24:43 2025] [::1]:52384 Accepted
-[Fri Jul  4 12:24:43 2025] [::1]:52384 Closing
-[Fri Jul  4 12:24:43 2025] [::1]:52386 Accepted
-[Fri Jul  4 12:24:43 2025] [::1]:52386 Closing
